+++ web-admin/tests/setup/fixtures/cli.ts
@@ -3,7 +3,7 @@ import {
   execAsync,
   spawnAndMatch,
   type SpawnAndMatchResult,
-} from "web-common/tests/utils/spawn";
+} from "@rilldata/web-common/tests/utils/spawn";
 
 export async function cliLogin(page: Page) {
   // Run the login command and capture the verification URL
