+++ web-local/tests/breadcrumbs.spec.ts
@@ -1,8 +1,10 @@
 import { expect } from "playwright/test";
-import { test } from "./utils/test";
+import { test } from "./setup/base";
 import { uploadFile } from "./utils/sourceHelpers";
 
 test.describe("Breadcrumbs", () => {
+  test.use({ project: { name: "Blank" } });
+
   test.describe("Breadcrumb interactions", () => {
     test("breadcrumb navigation", async ({ page }) => {
       await uploadFile(page, "AdBids.csv");
