+++ web-local/tests/data/projects/AdBids/models/AdBids_model.sql
@@ -1,0 +1,1 @@
+select * from AdBids
