+++ web-local/tests/data/projects/Blank/rill.yaml
@@ -1,0 +1,2 @@
+compiler: rill-beta
+title: "Test Project"
