+++ web-local/tests/metrics-editor.spec.ts
@@ -1,18 +1,15 @@
 import { expect } from "@playwright/test";
-import { useDashboardFlowTestSetup } from "web-local/tests/explores/dashboard-flow-test-setup";
-import {
-  AD_BIDS_EXPLORE_PATH,
-  AD_BIDS_METRICS_PATH,
-} from "web-local/tests/utils/dataSpecifcHelpers";
-import { gotoNavEntry } from "web-local/tests/utils/waitHelpers";
+import { gotoNavEntry } from "./utils/waitHelpers";
 import { updateCodeEditor, wrapRetryAssertion } from "./utils/commonHelpers";
-import { test } from "./utils/test";
+import { test } from "./setup/base";
 
 test.describe("Metrics editor", () => {
-  useDashboardFlowTestSetup();
+  test.use({ project: { name: "AdBids" } });
 
   test("Can add and remove measures and dimensions", async ({ page }) => {
-    await gotoNavEntry(page, AD_BIDS_METRICS_PATH);
+    await page.getByLabel("/metrics").click();
+    await page.getByLabel("/dashboards").click();
+    await gotoNavEntry(page, "/metrics/AdBids_metrics.yaml");
 
     await page.getByRole("button", { name: "Add new measure" }).click();
 
@@ -60,7 +57,9 @@ test.describe("Metrics editor", () => {
   });
 
   test("Metrics editor", async ({ page }) => {
-    await gotoNavEntry(page, AD_BIDS_METRICS_PATH);
+    await page.getByLabel("/metrics").click();
+    await page.getByLabel("/dashboards").click();
+    await gotoNavEntry(page, "/metrics/AdBids_metrics.yaml");
 
     await page.getByLabel("code").click();
 
@@ -78,11 +77,11 @@ test.describe("Metrics editor", () => {
     });
 
     // This is causing issues in the test, so we'll skip it for now.
-    // await gotoNavEntry(page, AD_BIDS_EXPLORE_PATH);
+    // await gotoNavEntry(page, "/dashboards/AdBids_metrics_explore.yaml");
     // // the Preview button should be disabled
     // await expect(page.getByRole("button", { name: "Preview" })).toBeDisabled();
     // await page.waitForTimeout(3000);
-    // await gotoNavEntry(page, AD_BIDS_METRICS_PATH);
+    // await gotoNavEntry(page, "/metrics/AdBids_metrics.yaml");
 
     // the editor should show a validation error
     await expect(
@@ -113,7 +112,7 @@ test.describe("Metrics editor", () => {
     await expect(page.getByText("Table columns")).toBeVisible();
 
     // go to the dashboard and make sure the metrics and dimensions are there.
-    await gotoNavEntry(page, AD_BIDS_EXPLORE_PATH);
+    await gotoNavEntry(page, "/dashboards/AdBids_metrics_explore.yaml");
 
     await page.getByRole("button", { name: "Preview" }).click();
